+++ src/main/java/com/openggf/level/objects/TouchActorContextPolicy.java
@@ -1,7 +1,8 @@
 package com.openggf.level.objects;
 
 public enum TouchActorContextPolicy {
-    MAIN_FULL_SIDEKICK_HURT_ONLY;
+    MAIN_FULL_SIDEKICK_HURT_ONLY,
+    MAIN_ONLY;
 
     public com.openggf.game.profiles.touchresponse.TouchActorContextPolicy toCanonical() {
         return com.openggf.game.profiles.touchresponse.TouchActorContextPolicy.valueOf(name());